size option for the dots in the scatter plot

+++ matplot.py
@@ -113,7 +113,7 @@ If a 'z' field is given, this field is used to color \
 the scatter dots. \
 If a 'e' field is given it is used to plot the error. \
 If --fields='*' is given all the columns are considered as y values.")
-scatter_options.add_option("-s", "--size", default=2, type=int,
+scatter_options.add_option("-s", "--size", default=2., type=float,
                            help='size of the dots for the scatter plot (default: 2)')
 parser.add_option("--labels", dest="labels", default=None, type='str',
                            help="Comma separated list of labels for each field \
